+++ DirectumRXAutoDeployer/Notifiers/Mattermost/Models/Response/Emoji.cs
@@ -5,8 +5,8 @@
         public string id { get; set; }
         public string creator_id { get; set; }
         public string name { get; set; }
-        public int create_at { get; set; }
-        public int update_at { get; set; }
-        public int delete_at { get; set; }
+        public long create_at { get; set; }
+        public long update_at { get; set; }
+        public long delete_at { get; set; }
     }
 }
